# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AnswerRoadDTO:
    properties:
      userCode:
        description: "ProgrammingID int    `json:\"programmindID\"`\n\tPathID        int
          \   `json:\"pathID\"`"
        type: string
    type: object
  dto.GetRoadDTO:
    properties:
      iconPath:
        type: string
      isFinished:
        type: boolean
      isStarted:
        type: boolean
      name:
        type: string
      paths:
        items:
          $ref: '#/definitions/dto.GetRoadPathDTO'
        type: array
    type: object
  dto.GetRoadPathDTO:
    properties:
      difficulty:
        type: integer
      id:
        type: integer
      isFinished:
        type: boolean
      isStarted:
        type: boolean
      languages:
        $ref: '#/definitions/dto.LanguageRoadDTO'
      name:
        type: string
    type: object
  dto.LanguageRoadDTO:
    properties:
      content:
        type: string
      description:
        type: string
      lang:
        type: string
      note:
        type: string
      title:
        type: string
    type: object
  dto.LogDTO:
    properties:
      content:
        type: string
      id:
        type: string
      labPathID:
        type: integer
      logType:
        type: string
      programmingID:
        type: integer
      userId:
        type: string
    type: object
...
  dto.PathDTO:
    properties:
      difficulty:
        type: integer
      id:
        type: integer
      isFinished:
        type: boolean
      isStarted:
        type: boolean
      language:
        $ref: '#/definitions/dto.LanguageRoadDTO'
      name:
        type: string
      template:
        type: string
    type: object
...
  response.BaseResponse:
    properties:
      data: {}
      data_count:
        type: integer
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost
info:
  contact: {}
  description: API Service for CodinLab
  title: API Service
  version: v1
paths:
...
  /private/log:
    get:
      consumes:
      - application/json
      description: Retrieves all logs based on the provided query parameters.
      parameters:
      - description: User ID
        in: query
        name: userID
        type: string
      - description: Programming ID
        in: query
        name: programmingID
        type: integer
      - description: Log Lab or Path ID
        in: query
        name: labRoadID
        type: integer
      - description: Log Content
        in: query
        name: content
        type: string
      - description: Log Type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LogDTO'
                  type: array
              type: object
      summary: Get all logs
      tags:
      - Log
...
  /private/log/rates:
    get:
      consumes:
      - application/json
      description: Retrieves language usage rates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Get Language Usage Rates
      tags:
      - Log
...
  /private/road/{programmingID}:
    get:
      consumes:
      - application/json
      description: Get Road with Paths
      parameters:
      - description: Language
        in: header
        name: Language
        type: string
      - description: programmingID
        in: path
        name: programmingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetRoadDTO'
              type: object
      summary: GetRoads
      tags:
      - Road
  /private/road/answer/{programmingID}/{pathID}:
    post:
      consumes:
      - application/json
      description: This is for answering quests.
      parameters:
      - description: Programming ID
        in: path
        name: programmingID
        required: true
        type: string
      - description: Path ID
        in: path
        name: pathID
        required: true
        type: string
      - description: Answer Road DTO
        in: body
        name: answerRoadDTO
        required: true
        schema:
          $ref: '#/definitions/dto.AnswerRoadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Answer
      tags:
      - Road
...
  /private/road/path/{programmingID}/{pathID}:
    get:
      consumes:
      - application/json
      description: Get Path By ID
      parameters:
      - description: Language
        in: header
        name: Language
        type: string
      - description: Programming ID
        in: path
        name: programmingID
        required: true
        type: string
      - description: Path ID
        in: path
        name: pathID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PathDTO'
              type: object
      summary: GetPathByID
      tags:
      - Road
...
  /private/road/reset/{programmingID}/{pathID}:
    get:
      consumes:
      - application/json
      description: Reset Path By Programming Language ID & Path ID
      parameters:
      - description: Path ID
        in: path
        name: pathID
        required: true
        type: string
      - description: Programming Language ID
        in: path
        name: programmingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: ResetPathHistory
      tags:
      - Road
  /private/start/{programmingID}:
    get:
      consumes:
      - application/json
      description: Start
      parameters:
      - description: programmingID
        in: path
        name: programmingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Start
      tags:
      - LabRoadCommon
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"